+++ packages/utils/src/schema/toPathSchema.ts
@@ -14,6 +14,7 @@ import {
   REF_KEY,
   RJSF_ADDITONAL_PROPERTIES_FLAG,
 } from '../constants';
+import getDiscriminatorFieldFromSchema from '../getDiscriminatorFieldFromSchema';
 import { FormContextType, PathSchema, RJSFSchema, StrictRJSFSchema, ValidatorType } from '../types';
 import getClosestMatchingOption from './getClosestMatchingOption';
 import retrieveSchema from './retrieveSchema';
@@ -52,20 +53,19 @@ function toPathSchemaInternal<T = any, S extends StrictRJSFSchema = RJSFSchema, 
     }
   }
 
-  const pathSchema: PathSchema = {
+  let pathSchema: PathSchema = {
     [NAME_KEY]: name.replace(/^\./, ''),
   } as PathSchema;
 
-  if (ONE_OF_KEY in schema) {
-    const index = getClosestMatchingOption<T, S, F>(validator, rootSchema!, formData, schema.oneOf as S[], 0);
-    const _schema: S = schema.oneOf![index] as S;
-    return toPathSchemaInternal<T, S, F>(validator, _schema, name, rootSchema, formData, _recurseList);
-  }
-
-  if (ANY_OF_KEY in schema) {
-    const index = getClosestMatchingOption<T, S, F>(validator, rootSchema!, formData, schema.anyOf as S[], 0);
-    const _schema: S = schema.anyOf![index] as S;
-    return toPathSchemaInternal<T, S, F>(validator, _schema, name, rootSchema, formData, _recurseList);
+  if (ONE_OF_KEY in schema || ANY_OF_KEY in schema) {
+    const xxxOf: S[] = ONE_OF_KEY in schema ? (schema.oneOf as S[]) : (schema.anyOf as S[]);
+    const discriminator = getDiscriminatorFieldFromSchema<S>(schema);
+    const index = getClosestMatchingOption<T, S, F>(validator, rootSchema!, formData, xxxOf, 0, discriminator);
+    const _schema: S = xxxOf![index] as S;
+    pathSchema = {
+      ...pathSchema,
+      ...toPathSchemaInternal<T, S, F>(validator, _schema, name, rootSchema, formData, _recurseList),
+    };
   }
 
   if (ADDITIONAL_PROPERTIES_KEY in schema && schema[ADDITIONAL_PROPERTIES_KEY] !== false) {